cleaned shader
